cloud_function_scripts/olist_customer_transformation/requirements.txt
functions-framework==3.8.1
google-cloud-storage==2.17.0
pandas==2.1.1
pyarrow==12.0.1
numpy==1.26.4